split config page in multiple components
